[Issue #862] update retina interfaces (#863)

+++ cpp/pixels-retina/test/RGVisibilityTest.cpp
@@ -70,8 +70,8 @@ TEST_F(RGVisibilityTest, BasicDeleteAndVisibility) {
 TEST_F(RGVisibilityTest, MultiThread) {
     struct DeleteRecord {
         uint64_t timestamp;
-        uint64_t rowId;
-        DeleteRecord(uint64_t timestamp, uint64_t rowId) : timestamp(timestamp), rowId(rowId) {}
+        uint32_t rowId;
+        DeleteRecord(uint64_t timestamp, uint32_t rowId) : timestamp(timestamp), rowId(rowId) {}
     };
 
     std::vector<DeleteRecord> deleteHistory;
@@ -133,15 +133,15 @@ TEST_F(RGVisibilityTest, MultiThread) {
         std::random_device rd;
         std::mt19937 gen(rd());
         
-        std::vector<uint64_t> remainingRows;
-        for (uint64_t i = 0; i < ROW_COUNT; i++) {
+        std::vector<uint32_t> remainingRows;
+        for (uint32_t i = 0; i < ROW_COUNT; i++) {
             remainingRows.push_back(i);
         }
 
         while (!remainingRows.empty() && running) {
             std::uniform_int_distribution<size_t> indexDist(0, remainingRows.size() - 1);
             size_t index = indexDist(gen);
-            uint64_t rowId = remainingRows[index];
+            uint32_t rowId = remainingRows[index];
 
             remainingRows[index] = remainingRows.back();
             remainingRows.pop_back();
